Add search
The search is executed entirely on the client site using a JSON file that contains the site's contents as index. All scripts and files are loaded lazily when the user activates the search.

The implementation is based on Arty2's 'fixedsearch.js' gist, although we use Flexsearch instead of Fuse. While everything has been tested, the code should probably be cleaned-up and refactored.