Simply handling of attributes for forms
This change keeps the order of the provided attributes in the final form. Unfortunately, this seems to be required for Netlify, which behaves weird if the 'netlify' attribute is not at the end of the <form> declaration. Additionally, this change simplifies specification of boolean attributes such as 'required'.

Existing form declaration will need to be updated. :(